# docs/swagger.yaml
definitions:
  constant.TaskStatus:
    enum:
    - Done!
    - Not done!
    - On going!
    - On planning!
    - Cancel!
    type: string
    x-enum-varnames:
    - TaskStatusDone
    - TaskStatusNotDone
    - TaskStatusOnGoing
    - TaskStatusOnPlanning
    - TaskStatusCancel
  entity.Task:
    properties:
      description:
        type: string
      end_date:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/constant.TaskStatus'
      task_id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  entity.User:
    properties:
      address:
        type: string
      dob:
        type: string
      password:
        type: string
      phone:
        type: string
      sex:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  request.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  wapper.Response:
    properties:
      data: {}
      error: {}
      status:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is Task-Management-System server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Task-Management-System API
  version: "1.0"
paths:
  /task/{user_id}:
    get:
      description: Get all tasks by user_id
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Get tasks by user_id
      tags:
      - task
  /task/available/{user_id}:
    get:
      description: Get all tasks end_time > now()
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Get tasks available
      tags:
      - task
  /task/create:
    post:
      consumes:
      - application/json
      description: Create task which user_id = ?
      parameters:
      - description: task create data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Create task
      tags:
      - task
  /task/expired/{user_id}:
    get:
      description: Get all tasks end_time < now()
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Get tasks expired
      tags:
      - task
  /task/update:
    put:
      consumes:
      - application/json
      description: Update task data
      parameters:
      - description: task update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Update task
      tags:
      - task
  /user/login:
    post:
      consumes:
      - application/json
      description: Login for user
      parameters:
      - description: user login data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Login user
      tags:
      - user
  /user/ping:
    get:
      description: get ping from service
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Ping service
      tags:
      - ping
  /user/register:
    post:
      consumes:
      - application/json
      description: Register for user
      parameters:
      - description: user register data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Register user
      tags:
      - user
  /user/update:
    put:
      consumes:
      - application/json
      description: Update infomations for user
      parameters:
      - description: user update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wapper.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/wapper.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wapper.Response'
      summary: Update user
      tags:
      - user
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"